changed module names, various comment changes, moved matplotlibcpp.h
