+++ README.md
@@ -10,7 +10,7 @@ network.
 
 If cloning the repo directly, you will need to download the C++ API from http://docs.hebi.us/downloads_changelogs.html#software to build the examples and place it in a folder called `hebi-cpp`. If you use the CMake project, it will automatically be downloaded, assuming you have a working internet connection.
 
-For examples which use plotting, matplotlibcpp is needed. Download matplotlibcpp.h from here: https://github.com/lava/matplotlib-cpp and place it in /hebi-cpp-examples/. In terminal, execute "sudo apt-get install python3-matplotlib python3-numpy python3.6-dev"
+For examples which use plotting, python is needed. To install it, execute "sudo apt-get install python3-matplotlib python3-numpy python3.6-dev" in terminal. (Plotting is enabled using a header file matplotlib.h which is a modified version of the one found at https://github.com/lava/matplotlib-cpp)
 
 Use the same version of the C++ API that is defined at the top of projects/cmake/DownloadHebiCpp.cmake, as this will ensure all the code is compatible.
