+++ basic/03a_command_position.cpp
@@ -72,6 +72,8 @@ int main() {
 
   // Stop logging
   auto log_file = group->stopLog();
+
+  //plot the logged position data
   std::vector<std::vector<double>> pos;
   pos.resize(group->size());
   GroupFeedback fbk(group->size());
