+++ basic/plot_functions.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "matplotlibcpp.h"
+#include "util/matplotlibcpp.h"
 
 namespace plt = matplotlibcpp;
